tests/wpt/tests/web-animations/responsive/shapeOutside.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id='container'>
    <div id='element'></div>
</div>
<script>

var container = document.getElementById('container');
var element = document.getElementById('element');

test(function() {
    container.style.fontSize = '10px';

    var keyframes = [
        {shapeOutside: 'circle(10em at 50% 50%)'},
        {shapeOutside: 'circle(10em at 50% 50%)'}
    ];
    var player = element.animate(keyframes, 10);
    player.pause();
    player.currentTime = 5;

    var shapeOutside = getComputedStyle(element).shapeOutside;
    container.style.fontSize = '20px';
    assert_not_equals(getComputedStyle(element).shapeOutside, shapeOutside);
}, 'shapeOutside responsive to style changes');

test(function() {
    var keyframes = [
        {shapeOutside: 'inherit'},
        {shapeOutside: 'circle(200px at 50% 50%)'}
    ];

    container.style.shapeOutside = 'circle(100px at 50% 50%)';
    var player = element.animate(keyframes, 10);

    player.pause();
    player.currentTime = 5;
    assert_equals(getComputedStyle(element).shapeOutside, 'circle(150px at 50% 50%)');

    container.style.shapeOutside = 'inset(100%)';
    assert_equals(getComputedStyle(element).shapeOutside, 'circle(200px at 50% 50%)');

    container.style.shapeOutside = 'circle(400px at 50% 50%)';
    assert_equals(getComputedStyle(element).shapeOutside, 'circle(300px at 50% 50%)');

    container.style.shapeOutside = 'circle(800px at 50% 50%)';
    assert_equals(getComputedStyle(element).shapeOutside, 'circle(500px at 50% 50%)');

    container.style.shapeOutside = 'inset(100%)';
    assert_equals(getComputedStyle(element).shapeOutside, 'circle(200px at 50% 50%)');

    container.style.shapeOutside = 'circle(800px at 50% 50%)';
    assert_equals(getComputedStyle(element).shapeOutside, 'circle(500px at 50% 50%)');

    container.style.shapeOutside = 'circle(400px at 50% 50%)';
    assert_equals(getComputedStyle(element).shapeOutside, 'circle(300px at 50% 50%)');
}, 'shapeOutside responsive to inherited shapeOutside changes');

test(() => {
    const target = document.createElement('div');
    document.body.appendChild(target);
    target.style.shapeOutside = 'path("M 10 10 L 20 20 Z") border-box';

    const animation = target.animate([
        {},
        {shapeOutside: 'path("M 50 50 L 60 60 Z") border-box'},
    ], {
        duration: 1000,
        fill: 'both',
    });
    animation.pause();
    animation.currentTime = 500;

    assert_equals(getComputedStyle(target).shapeOutside,
                  'path("M 30 30 L 40 40 Z") border-box');

    target.style.shapeOutside = 'path("M 10 10 L 20 20 Z") padding-box';
    assert_equals(getComputedStyle(target).shapeOutside,
                  'path("M 50 50 L 60 60 Z") border-box');

    animation.cancel();
    target.remove();
}, 'Neutral path() keyframes invalidate when the underlying <shape-box> changes');
</script>
