LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/shapeOutside-expected.txt


PASS shapeOutside responsive to style changes
FAIL shapeOutside responsive to inherited shapeOutside changes assert_equals: expected "circle(200px at 50% 50%)" but got "circle(150px at 50% 50%)"
